+++ packages/eyes-cypress/src/plugin/pluginExport.js
@@ -1,6 +1,6 @@
 'use strict';
 const setGlobalRunHooks = require('./hooks');
-const CYPRESS_SUPPORTED_VERSION = '6.2.0';
+const shouldSetGlobalHooks = require('./shouldSetGlobalHooks');
 
 function makePluginExport({startServer, eyesConfig, visualGridClient, logger}) {
   return function pluginExport(pluginModule) {
@@ -12,9 +12,8 @@ function makePluginExport({startServer, eyesConfig, visualGridClient, logger}) {
       const moduleExportsResult = await pluginModuleExports(...args);
       const [on, config] = args;
 
-      if (config.version >= CYPRESS_SUPPORTED_VERSION && config.experimentalRunEvents) {
-        setGlobalRunHooks(on, {visualGridClient, logger});
-        eyesConfig.eyesLegacyHooks = false;
+      if (shouldSetGlobalHooks({...config, ...eyesConfig})) {
+        setGlobalRunHooks(on, {visualGridClient, logger, eyesConfig});
       }
 
       return Object.assign({}, eyesConfig, {eyesPort}, moduleExportsResult);