+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- remove the need for `experimentalRunEvents` in cypress >= `6.7.0`
 
 ## 3.21.0 - 2021/4/22
 
@@ -9,7 +10,7 @@
 - updated to @applitools/visual-grid-client@15.8.0 (from 15.6.2)
 
 ## 3.20.4 - 2021/3/17
-
+s
 - set explicit timeout for eyesCheckWindow [trello](https://trello.com/c/5Thz1QFs)
 - updated to @applitools/dom-snapshot@4.4.14 (from 4.4.13)
 - updated to @applitools/visual-grid-client@15.6.2 (from 15.6.0)